requirements.txt
pillow
streamlit>=0.82.0
requests